refactor typeahead
