+++ src/components/input/_input.scss
@@ -103,3 +103,82 @@
     color: $color-red;
   }
 }
+
+.typeahead {
+  position: relative;
+
+  &__combobox {
+    display: inline-block;
+    border-radius: $input-radius;
+
+    @include mq('xs', 's') {
+      width: 100%;
+    }
+  }
+
+  &__results {
+    display: none;
+    margin: 0.5rem 0 0;
+    padding: 0;
+    width: 100%;
+
+    overflow: hidden;
+
+    border: 1px solid $color-input;
+    border-radius: $input-radius;
+  }
+
+  &__results-title {
+    padding: 0.25rem 0.5rem;
+    border-bottom: 1px solid $color-input;
+    background: $color-grey-4;
+  }
+
+  &__listbox {
+    margin: 0;
+    padding: 0;
+    list-style: none;
+    background: $color-white;
+  }
+
+  &__option {
+    margin: 0;
+    padding: $input-padding-horizontal;
+    outline: none;
+
+    cursor: pointer;
+
+    &:not(:last-child) {
+      border-bottom: 1px solid $color-input;
+    }
+
+    &:not(&--no-results):not(&--more-results):hover,
+    &--focused:not(&--no-results) {
+      border-color: $color-blue;
+      background: $color-blue;
+      color: $color-white;
+    }
+
+    &--no-results,
+    &--more-results {
+      padding: 0.25rem 0.5rem;
+      background: $color-grey-4;
+      cursor: not-allowed;
+    }
+  }
+
+  // Modifiers
+  &:not(&--initialised) & {
+    &__instructions,
+    &__listbox,
+    &__status {
+      display: none;
+    }
+  }
+
+  &--has-results & {
+    &__results {
+      display: block;
+    }
+  }
+}
