+++ src/components/input/_macro-options.md
@@ -20,6 +20,7 @@
 | legend            | string                                                        | Only if mutuallyExclusive is set | Text content for the legend                                                         |
 | legendClasses     | string                                                        | false                            | Classes for the legend                                                              |
 | error             | `Error` [_(ref)_](/components/error)                          | false                            | Configuration for validation errors                                                 |
+| typaheadData      | string                                                        | false                            | JSON file path for typeahead                                                        |
 
 ## Prefix/Suffix
 
@@ -28,3 +29,17 @@
 | text  | string | true     | The text for the prefix/suffix                                                                   |
 | title | string | true     | The title of the prefix/suffix. For example where `text` is "cm", `title` would be "centimeters" |
 | id    | string | false    | Id for the prefix/suffix                                                                         |
+
+## Typeahead
+
+| Name          | Type                                 | Required | Description                                                              |
+| ------------- | ------------------------------------ | -------- | ------------------------------------------------------------------------ |
+| id            | string                               | true     | ID for the input                                                         |
+| classes       | string                               | false    | CSS classes to apply to the wrapping element                             |
+| content       | TypeaheadContent                     | true     | Aria and results messaging content                                       |
+| typeaheadData | string                               | true     | URL of the JSON file with the typeahead data that needs to be searched   |
+| inputClasses  | string                               | false    | CSS classes to apply to the input                                        |
+| name          | string                               | true     | Name attribute for the input                                             |
+| label         | `Label` [_(ref)_](/components/label) | true     | Label config for the input                                               |
+| autocomplete  | string                               | true     | Autocomplete attribute used to override the browsers native autocomplete |
+| error         | `Error` [_(ref)_](/components/error) | false    | Configuration for validation errors                                      |
