+++ src/components/input/examples/country-of-birth/index.njk
@@ -1,13 +1,10 @@
----
-lang: cy
----
-{% from "components/typeahead/_macro.njk" import onsTypeahead %}
+{% from "components/input/_macro.njk" import onsInput %}
 
-{{ onsTypeahead({
+{{ onsInput({
     "id": "country-of-birth",
     "label": {
-        "text": "Nodwch enw y wlad",
-        "description": "Nodwch eich ateb neu dewiswch o'r awgrymiadau"
+        "text": "Current name of country",
+        "description": "Enter your own answer or select from suggestions"
     },
     "instructions": "Use up and down keys to navigate suggestions once you\'ve typed more than two characters. Use the enter key to select a suggestion. Touch device users, explore by touch or with swipe gestures.",
     "content": {
