--- src/components/typeahead/_test-template.njk
@@ -1,3 +1,0 @@
-{% from "components/typeahead/_macro.njk" import onsTypeahead %}
-
-{{ onsTypeahead(params) }}
