--- src/components/typeahead/index.njk
@@ -1,60 +1,0 @@
----
-title: Typeahead
----
-{% from "views/partials/example/_macro.njk" import patternlibExample %}
-{% from "components/panel/_macro.njk" import onsPanel %}
-
-Typeaheads are used to suggest options for answers and typically search code lists such as, Country of Birth, Languages, Nationalities, etc.
-
-{{
-    patternlibExample({"path": "components/typeahead/examples/country-of-birth/index.njk"})
-}}
-
-## How to use this component
-
-These examples are powered by a JavaScript library called [Fuse.js](https://fusejs.io/). 
-
-### Response object
-
-The typeaheads expect the response from the library to return the results as JSON objects in an array.
-
-### TypeaheadResultItem
-
-The TypeaheadResultItem should consist of at least a key corresponding to the searched language (e.g. `en-gb`) with its corresponding value.
-
-### Example response
-
-```
-{
-    "results": [
-        {
-            "en-gb": "United Kingdom",
-            "cy": "Deyrnas Unedig"
-        }
-    ],
-    "totalResults": 1
-}
-```
-
-## Variants
-
-### Country of birth (Cymraeg)
-
-The typeahead component sends a `searchFields` parameter in it's query string that matches the `lang` attribtue on the `<html>` element of the page.
-
-The typeahead component will search the `cy` or the `en-gb` data to find results depending on the `searchFields` parameter.
-
-{{
-    patternlibExample({"path": "components/typeahead/examples/country-of-birth-cy/index.njk"})
-}}
-
-## Research on this component
-
-{% call  onsPanel() %}
-    <p>If you have conducted any user research using this component, please feed back your findings via the <a href="https://github.com/ONSdigital/design-system/issues/566">Design System Forum.</a></p>
-{% endcall %}
-
-## Design System forum
-
-[Discuss 'Typeahead' on GitHub](https://github.com/ONSdigital/design-system/issues/566)
-
